+++ core/src/integrationTest/java/tech/pegasys/web3signer/core/jsonrpcproxy/support/TestEth1Config.java
@@ -16,6 +16,8 @@ import tech.pegasys.web3signer.core.config.Eth1Config;
 import tech.pegasys.web3signer.core.config.client.ClientTlsOptions;
 import tech.pegasys.web3signer.core.service.jsonrpc.handlers.signing.ChainIdProvider;
 import tech.pegasys.web3signer.core.service.jsonrpc.handlers.signing.ConfigurationChainId;
+import tech.pegasys.web3signer.signing.config.AzureKeyVaultParameters;
+import tech.pegasys.web3signer.signing.config.DefaultAzureKeyVaultParameters;
 
 import java.time.Duration;
 import java.util.Optional;
@@ -88,5 +90,10 @@ public class TestEth1Config implements Eth1Config {
   @Override
   public ChainIdProvider getChainId() {
     return chainId;
+  }
+
+  @Override
+  public AzureKeyVaultParameters getAzureKeyVaultConfig() {
+    return new DefaultAzureKeyVaultParameters("", "", "", "");
   }
 }